Fix test header include paths Rename `gpu/test_tools.cpp` to `gpu/gpu_test_tools.cpp` for better organization
