+++ tests/gpu/gpu_test_tools.cpp
@@ -1,4 +1,4 @@
-#include "test_tools.h"
+#include "gpu_test_tools.h"
 
 #include <exception>
 #include <iostream>
