+++ tests/linalg_test/Det_test.cpp
@@ -4,7 +4,7 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
-#include "../test_tools.h"
+#include "test_tools.h"
 #include "Device.hpp"
 #include "linalg.hpp"
 #include "Type.hpp"
